Encoding Zip Code
Program takes a five-digit zip code and returns a bar code according to the encoding scheme used by the United States Postal Service. If the sum of the digits is not a multiple of 10, the sixth frame bar is set to 1. The digit is computed from the bar code using the column weights 7, 4, 2, 1, 0, each of which is multiplied by the 0 or 1 for that bar and added together, such as 1 x 7 + 1 x 4 + 0 x 2 + 0 x 1 + 0 x 0 = 11. Then it creates a string of 0's and 1's corresponding to each digit, and prints the bar code representation of each digit of the zip code.