Discount for Pet Store Customer
Program prompts the user to enter each price of a sale then a Y for a pet or N for another item. A discount equal to 20 percent of the cost of items in a sale other than the pets is added if the customer buys one or more pets and at least five other items.